Update README.md
Heritability of a trait is often identified and reported in an ancestry group stratified manner. This limits the ability to estimate and report the combined heritability in a multi-ethnic population. Although there are several methods demonstrated recently with robust ways of calculating heritability with or without individual-level datasets, these methods are limited to ancestry-specific groups.  In this project, we are proposing a way to estimate combined heritability using a multi-ethnic reference linkage-disequilibrium (LD) panel with equal proportions of data. We will use current existing tools to simulate and calculate heritability and report it as a framework that can be implemented and explored further. This will lead to developing a novel approach to estimating the heritability of the particular trait (NTproBNP, Blood Pressure) in multi-ethnic populations. As a part of Team HeriVar, you will be contributing to the demonstration of methodology, and heritability calculation, and working as a team to promote the method.